requirements.txt
opencv-python
numpy
pynetworktables
ultralytics
scytheautoupdate
Flask
torch